implemented checking of "freshness" of the downloaded posts

+++ app/src/main/java/ru/home/customvk/PostRepository.kt
@@ -34,8 +34,10 @@ private class DefaultPostRepository : PostRepository {
         }
 
     private fun downloadAndSavePosts(): Single<List<Post>> = fetchPostsFromInternet().map { downloadedPosts ->
-        postDao.deleteAllPosts()
-        postDao.savePosts(downloadedPosts)
+        if (downloadedPosts.isNotEmpty()) {
+            postDao.deleteAllPosts()
+            postDao.savePosts(downloadedPosts)
+        }
         downloadedPosts
     }
